requirements.txt
discord==2.3.2
discord.py==2.3.2
python-dotenv==1.0.1
jsonschema==4.6.0